backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
firebase-admin==6.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0
google-generativeai==0.3.2
langchain==0.0.350
langchain-google-genai==0.0.6
aiofiles==23.2.1
httpx==0.25.2
cryptography==41.0.7
bcrypt==4.1.2
slowapi==0.1.9